pyproject.toml
[tool.poetry]
name = "dtpr"
version = "0.1.0"
description = "Set of tools to implement patter recognition algorithms on CMS DTs"
authors = ["Daniel Estrada <estradadaniel@uniovi.es>"]
readme = "README.md"

[tool.poetry.dependencies]
# You need to install also ROOT
python = ">=3.9"
numpy = ">=2.0"
pandas = ">=2.2"
matplotlib = ">=3.9"
mplhep = ">=0.3"
PyYAML = ">=5.4.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "*"

[tool.poetry.group.doc]
optional = true

[tool.poetry.group.doc.dependencies]
sphinx = "*"
sphinx-rtd-theme = "*"
sphinx-autobuild = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"